+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 - [Feature] Add ability to specify a left or right icon for **Input** ([#1320](https://github.com/optimizely/oui/pull/1320))
 - [Feature] Add ability to include a Clear Input ButtonIcon with `hasClearButton` for **Input** ([#1320](https://github.com/optimizely/oui/pull/1320))
 - [Feature] Add ability to remove bottom border and box shadow with `toolbarStyle` in **Toolbar** ([#1327](https://github.com/optimizely/oui/pull/1327))
+- [Patch] Remove mockdate from stories to fix Storybook knobs ([#1329](https://github.com/optimizely/oui/pull/1329))
 
 ## 46.0.4 - 2020-04-30
 - [Patch] Add search function to the **SearchPicker** render prop API ([1323](https://github.com/optimizely/oui/pull/1323))